Added instructions for creating backend config.
At step4 of the install routine, there were instructions
for generating a json db config file for the frontend.

However, there was no hint how to give db access to
the backend.

Now there are instructions how to create a backend-readable
config file (it must be named db.cfg for some reason).

NOTE: Future implementations should give a json parser
to the backend and alter the required filenames, so
that only one config file is needed.




git-svn-id: https://devel.etoa.net/svn/etoa-gui/trunk@1474 9ddb851b-07ab-4ff2-becd-2ceea2d3057d