#99 improved checks for running research and defense
APPLY DB PATCH FROM [1477/EtoA-GUI]
BEFORE UPDATING TO THIS COMMIT!

Added a field to def_query to save the time when the
user actually clicked a defensive fortification.
(= the time when it was checked whether there is a
corresponding specialist for fewer ressources)

Now everything should work as expected:
The specialists are only blocked if something is
building what has been started at a time when the
specialist was already there.



git-svn-id: https://devel.etoa.net/svn/etoa-gui/trunk@1478 9ddb851b-07ab-4ff2-becd-2ceea2d3057d