#102 solved: changed calc function in display and added ceil() wrappers.
This is just a dirty fix.
The function still appears at two places:
Once in alliance.class.php and once in
base.inc.php - same code twice.

Added hint comments, but this should be
rewritten correctly.

Also added ceil() wrappers, because we do not
want to have a non-integer number of ship parts.




git-svn-id: https://devel.etoa.net/svn/etoa-gui/trunk@1487 9ddb851b-07ab-4ff2-becd-2ceea2d3057d