Removes problems with tech stealing
The spyattack_count property stores only the
number of successful spyattacks now.

So there shouldn't be such a fast decline in
chance like there was with the new techsteal
algoritm previously.

Also if a user sends a mysticum along the spy
ships that has a 100% chance of stealing
researched, it will steal a tech for sure,
not only with a probability of max_chance like
before.

There is still the -0.2% chance times the square
root of the number of successful spy attacks.
So with 226 successful spy attacks the user has
reached a below-zero value for the chance if
he sends only one ship.