removes problems with alliance-Wing member switching
Number of member for which an alliance building
was built for did decrease in some cases
of member switching / adding.

Now the query only changes it if the new number of
members the building is built for increases.

This only affects member adding / removing /
switching. If a building increases its level,
the member number is checked again and also
still decreased if necessary.



git-svn-id: https://devel.etoa.net/svn/etoa-gui/trunk@1491 9ddb851b-07ab-4ff2-becd-2ceea2d3057d