#98 added keybinds for faster navigation
This commit includes several changes to provide
a basic keyboard navigation for etoa.

There is a new config value for users to turn
keybinds on or off. Be sure to apply the db patch.

At the moment, accessing the most important pages
and switching between prev/next planet is possible.

New keys or changes can be added in the keybinds.js.

keybinds.js has the strange behaviour of always
generating a http request for it on every new
loading page, the server responds with a 304 Not
Modified answer code.




git-svn-id: https://devel.etoa.net/svn/etoa-gui/trunk@1528 9ddb851b-07ab-4ff2-becd-2ceea2d3057d