Fixed MySQL injection vulnerability in haven target form
Previously a string in one of the form inputs "man_sx",
"man_sy" etc. like
1";DROP TABLE `users`; --
passed the checks (because php only looks at the 1
at the beginning and takes this as the numeric
value, later the whole string is inserted into the
query string.

Now it only takes the intval() values for the checks
and the query string.




git-svn-id: https://devel.etoa.net/svn/etoa-gui/trunk@1560 9ddb851b-07ab-4ff2-becd-2ceea2d3057d