#873 moved citybox css from admin to general, added autofocus
In [1619/EtoA-GUI] I accidentially added a citybox div
to htdocs/content/messages/new.php which made the 
autocomplete work again but it now looked broken.

This is because the citybox-related css was only in the
admin/default.css - so that is now moved to general.css

When a name was autocompleted, the name input field
lost focus. So now I added a xajax line that gives
calls focus() on this input after autocompletion.




git-svn-id: https://devel.etoa.net/svn/etoa-gui/trunk@1621 9ddb851b-07ab-4ff2-becd-2ceea2d3057d