Improved chat: added mutex for update(), #870 improved scrolling detection, added new messages indicator
The update() function is now inside a mutex to avoid double messages.
Because javascript has no support for atomics, locks etc.,
Lamport's bakery algorithm is used for mutual exclusion.

This should resolve the problem with twice shown messages.

To avoid wrong results when checking whether the user has scrolled
way down to the bottom, the property "clientHeight" instead of
"outerHeight" is now used. Also the check is now for less-than,
because sometimes a negative value may result in some browsers.

A small box is added where the number of unread messages (unread
= below the visible part because not autoscrolling) is shown.
If the scroll detection fails, the user now knows there are new
messages he hasn't read.

When scrolling down, the shown number of unread messages is reduced
and the indicator box is completely removed if the bottom is reached.
If the user scrolls up again, the number of unread messages stays
the same.

The chat "login" command now works like the "update" command
to avoid double-updating issues; chat messages are already
sent with the login json object, so no second poll() is needed.




git-svn-id: https://devel.etoa.net/svn/etoa-gui/trunk@1624 9ddb851b-07ab-4ff2-becd-2ceea2d3057d