Autocomplete for usernames now ignores leading spaces
Many users that want to send a message to multiple
recipients insert a ; after the first name but also
a space after the ; - so autocomplete fails and they
get an error message (user x does not exist), because
there is no user whose name starts with a space.

Now autocomplete trims spaces and removes leading
spaces from autocompleted names.



git-svn-id: https://devel.etoa.net/svn/etoa-gui/trunk@1629 9ddb851b-07ab-4ff2-becd-2ceea2d3057d